+++ src/rex.formbuilder/static/js/lib/elements/questions/Boolean.js
@@ -5,8 +5,11 @@
 'use strict';
 
 var objectPath = require('object-path');
+var deepCopy = require('deep-copy');
 
 var Question = require('./Question');
+var properties = require('../../properties');
+var {isEmpty} = require('../../util');
 var _ = require('../../i18n').gettext;
 
 
@@ -19,6 +22,60 @@ class BooleanQuestion extends Question {
     return 'question-boolean';
   }
 
+  static getPropertyConfiguration(isSubElement) {
+    var cfg = Question.getPropertyConfiguration(isSubElement);
+    cfg.properties.advanced.unshift(
+      {
+        name: 'length',
+        minLabel: _('Minimum Length'),
+        maxLabel: _('Maximum Length'),
+        schema: properties.NumericRange
+      },
+      {
+        name: 'pattern',
+        label: _('Pattern Constraint'),
+        schema: properties.Regex
+      },
+      {
+        name: 'width',
+        label: _('Field Width'),
+        schema: properties.WidgetSize
+      },
+      {
+        name: 'hotkeys',
+        label: _('Hotkeys'),
+        schema: properties.BooleanHotkeys
+      }
+    );
+    return cfg;
+  }
+
+  constructor() {
+    super();
+    this.hotkeys = {};
+  }
+
+  parse(element, instrument, field) {
+    super.parse(element, instrument, field);
+    var yes = objectPath.get(
+      element,
+      'options.widget.options.hotkeys.true',
+      null
+    );
+    if (yes !== null) {
+      this.hotkeys.yes = yes;
+    }
+
+    var no = objectPath.get(
+      element,
+      'options.widget.options.hotkeys.false',
+      null
+    );
+    if (no !== null) {
+      this.hotkeys.no = no;
+    }
+  }
+
   serialize(instrument, form, context) {
     context = context || this;
 
@@ -28,10 +85,36 @@ class BooleanQuestion extends Question {
     var field = context.getCurrentSerializationField(instrument);
     field.type = 'boolean';
 
+    if (!isEmpty(this.hotkeys)) {
+      var elm = context.getCurrentSerializationElement(form);
+      objectPath.set(elm, 'options.widget.type', 'radioGroup');
+
+      if (this.hotkeys.yes) {
+        objectPath.set(
+          elm,
+          'options.widget.options.hotkeys.true',
+          this.hotkeys.yes
+        );
+      }
+      if (this.hotkeys.no) {
+        objectPath.set(
+          elm,
+          'options.widget.options.hotkeys.false',
+          this.hotkeys.no
+        );
+      }
+    }
+
     return {
       instrument,
       form
     };
+  }
+
+  clone(exact, configurationScope) {
+    var newElm = super.clone(exact, configurationScope);
+    newElm.hotkeys = deepCopy(this.hotkeys);
+    return newElm;
   }
 }